requirements.txt
aiohttp
discord.py
discord.py[voice]
gtts
langdetect
python-dotenv
pyttsx3
PyYAML
Requests
openai
duckduckgo-search
flask
aiohttp
discord.py
duckduckgo-search
flask
gtts
langdetect
openai
python-dotenv
pyttsx3
PyYAML
requests
click
aiohttp
click
discord.py
duckduckgo-search
flask
gtts
langdetect
openai
python-dotenv
pyttsx3
PyYAML
requests